Final Changes

+++ db/migrate/20210902103716_add_salary_to_user.rb
@@ -1,5 +1,5 @@
 class AddSalaryToUser < ActiveRecord::Migration[6.0]
   def change
-    add_column :users, :salary, :integer
+    add_column :users, :salary, :float
   end
 end
